requirements.txt
fastAPI
pytest
uvicorn
python-dotenv
pytest-cov
pylint
black
pika
pytest-asyncio
pytest-integration
pytest-envvars
asyncio
argon2-cffi
python-jose[cryptography]
logging-test-case
firebase-admin
pydantic-settings
pyrebase4